Make call to Compute_Overhead_O3 in toms_mod.F90 more robust
This commit fixes the issue reported in geoschem/geos-chem #591.

On some older compilers, an error is caused because State_Chm%TO3_DAILY
was not declared with INTENT(IN) instead of INTENT(INOUT).  Newer
compilers seem to get around this issue (see explanation on the Github
issue #591 page) but some older compilers (like ifort 12) do not.

To implement a robust solution, we now explicitly pass State_Chm to
Compute_Overhead_O3 in toms_mod.F90 via a dummy argument that is
declared with INTENT(INOUT).  This should be agnostic of compiler version.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>