Remove Emissions Menu from input.geos and now obtain Input_Opt%DoEmissions from HEMCO
In the Emissions Menu of input.geos we currently have:

  %%% EMISSIONS MENU %%%  :
  HEMCO Input file        : HEMCO_Config.rc
  Switches for UCX        :---
   => Use CH4 emissions?  : F
   => Set init. strat. H2O: T

The switch for Set init. strat. H2O has been a source of user confusion
because this option should always be turned off for runs other than an
initial simulation but often users forget (or don't know) to update this
for subsequent runs. We will remove this switch and if users would like to
change the initial concentrations of H2O for their simulations, they should
do so by manipulating the initial conditions (i.e. restart file) manually.

The switch for Use CH4 emissions is not needed. We should rely completely
on settings in HEMCO_Config.rc for emissions. This switch enables calls to
routine EMISSCH4 in global_ch4_mod.F90 but that routine should only be used
for CH4-only simulations, and is also only needed in those simulations for
perturbing emissions in the case of analytical inversions. Ideally this
would also be done in HEMCO_Config.rc in the future.

Finally, the setting for HEMCO Input file : HEMCO_Config.rc should really
not be changed. We don't allow users to modify the filenames for other
input files (input.geos, ExtData.rc, HISTORY.rc), so we should also have
HEMCO_Config.rc as the standard name for the HEMCO configuration file. Users
can always have multiple versions of this file with different names if they
want to evaluate and save different emissions settings.

The emissions switch is now controlled by the EMISSIONS setting in
HEMCO_Config.rc. We now create a shadow field in GEOS-Chem (Input_Opt%DoEmissions)
to determine whether we need to apply emissions flux in mixing_mod.F90.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>