Use sed commands that will work on MacOS and Linux in setupConfigFiles.sh
This commit resolves Github issue geoschem/geos-chem #617.  The MacOS
version of the "sed" command (used for replacing text in files)
differs slightly from the GNU/Linux version.

To fix this issue, we have created a wrapper routine sed_ie in
the script run/shared/setupConfigFile.sh.  The sed_ie routine executes
the "sed -i -e" command in such a way that it will work on both MacOS
and GNU/Linux.  We have updated other calls to "sed -i -e" in this
script to use the wrapper function "sed_ie".

There may be additional instances that need modification, but this
should be sufficient to define most GEOS-Chem-Classic run directories.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>