Bug fix: For tagCH4 sims, do not test if CH4_SAB species is negative
In the tagCH4 simulation, an error occurs in routine DO_TEND in
mixing_mod.F90.  The problem is that the CH4_SAB species (which is
a soil absorption and can be either positive or negative) trips an
error check for negatives.

The simple fix is to skip flagging negative CH4_SAB species as an
error.  The warning message is suppressed and the run is allowed
to continue.

NOTE: The tagCH4 simulation is not used regularly as of Feb 2021.
But this issue was causing the tagCH4 simulation to fail integration
tests.  We defer to the Carbon Cycle Working Group for a more
permanent solution.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>