Add fixes for interpreting the the ObsPack sampling window times
This commit fixes the issue that Andy Jacobson brought up on
Github issue geoschem/geos-chem #631.  The "time" variable indicates
the start of the sampling window unless specified otherwise in
the time:comment attribute.

We now assume that "time" indicates the start of the sampling interval
if the time:comment attribute is not found.  If it is found, we use
parse text to see if "time" indicates the start time or midpoint time
of the start window.  We then compute the State_Diag%ObsPack_Ival_Start,
State_Diag%ObsPack_Ival_Center, and State_Diag%ObsPack_Ival_End
arrays accordingly.

Also trimmed trailing whitespace in ObsPack/obspack_mod.F90

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>