Update default run directory name to include met field
Users should create new run directories depending on if they set up for
GEOS-FP or MERRA-2 simulations. Depending on which met field is chosen,
the native resolution uses for offline emissions files will change. To
make it clear which met fields should be used with the run directory, we
now include that string in the default run directory name.

Users don't need to create new directories if they want to change model
resolution, for either GCClassic or GCHP so we don't need to include this
string in the default run directory name.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>