Rundir fixes: Copy restart file when on AWS; Fix typo in fullchem rst name
Added code to check the Linux Kernel release (on AWS, it always has
"-aws" in the name).  If that is found, then use the aws s3 cp
command to copy the sample restart file to the run directory,
(GEOS-Chem Classic only)

Also a Git merge probably clobbered the date of the default fullchem
restart file.  This should have been 20190701 but was 20160701instead.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>