Add int tests for aerosol-only and TOMAS40; remove 2x25_fullchem_benchmark
The aerosol-only simulation in GC_13.0.0 is now functional once more,
so we have added an integration tests for it.

The TOMAS40 simulation now uses a proper restart file, so we can
also schedule integration tests for this as well.

The 2x25_fullchem_benchmark integration tests will fail because the mass
tuning factors for dust are only provided at 4x5.  Therefore we will
comment out the code that creates 2x25_fullchem_benchmark rundirs.

gned-off-by: Bob Yantosca <yantosca@seas.harvard.edu>