Add fix to avoid "Is_Gas and Is_Aerosol are both false" error when adding new P/L species in KPP
If a species is not defined in the species database, it gets assigned Is_Gas=FALSE
and Is_Aerosol=FALSE, triggering an error trap in species_database_mod.F90.
When new prod/loss families or dummy species are added in KPP, users don't
always want to add them to the species database. The check for Is_Gas and
Is_Aerosol set to false has been updated to further check if the species
Is_Kpp, and if so to set Is_Gas=TRUE to avoid the model crashing.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>