Update intTestExecute_slurm.sh to ensure OMP_NUM_THREADS matches number of CPUs requested by SLURM
This fix avoids an issue with output printing out twice to the log files when
running on Harvard's huce_cascade partition on the Cannon cluster. This happens
when the number of OMP threads is less than the number of CPUs requested by
SLURM. Eventually the duplicate run will crash by hitting the memory limit
causing a failed job code and resulting in the integration test failure.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>